Move WitsmlBitRecord and WitsmlNameTag one directory up, fix TubularTests tubular file path
