+++ Src/Witsml/Data/WitsmlBitRecord.cs
@@ -2,7 +2,7 @@ using System.Xml;
 using System.Xml.Serialization;
 using Witsml.Data.Measures;
 
-namespace Witsml.Data.Tubular
+namespace Witsml.Data
 {
     public class WitsmlBitRecord
     {
